+++ packages/types/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.31] - 2026-04-23
+
+- Add Element properties to support Advanced number inputs
+
 ## [1.0.30] - 2026-03-30
 
 - Add ValidationInputType as sub-type of HTMLTextInputTypeAttribute